added multirecord

+++ modules/blob.py
@@ -21,29 +21,39 @@ class BlobUpload(models.Model):
     storage_account_url = fields.Char(required=True)
     container = fields.Char(required=True)
     blob_name = fields.Char(required=True)
-    credential = fields.Char(requried=True)
+    credential = fields.Char(required=True)
 
-    def _do_upload(self, blob):
-        with open(self.file, "rb") as data:
+    def _do_upload(self, blob, file_full_path):
+        with open(file_full_path, "rb") as data:
             blob.upload_blob(data)
-        _logger.info("Local file {} uploaded to Blob Storage: {}".format(self.file, self.blob_name))
+        _logger.info("Local file {} uploaded to Blob Storage: {}".format(file_full_path, self.blob_name))
 
     def upload(self, remove_local_data_file=True, overwrite=True):
+        for record in self:
+            try:
+                file_full_path = os.path.join(config.get('data_dir'), record.file)
+                if not os.path.exists(file_full_path):
+                    _logger.error("Cannot find file ({}), thus not uploading".format(file_full_path))
+                    continue
 
-        blob = BlobClient(
-            account_url=self.storage_account_url,
-            container_name=self.container,
-            blob_name=self.blob_name,
-            credential=self.credential)
+                blob = BlobClient(
+                    account_url=record.storage_account_url,
+                    container_name=record.container,
+                    blob_name=record.blob_name,
+                    credential=record.credential)
 
-        try:
-            self._do_upload(blob)
-        except ResourceExistsError as ex:
-            if not overwrite: 
-                raise ex
-            _logger.debug("El fichero existe, hay que sobreescribirlo")
-            blob.delete_blob()
-            _logger.debug("El fichero se ha borrado para sobreescribirlo")
-            self._do_upload(blob)
+                try:
+                    record._do_upload(blob, file_full_path)
+                except ResourceExistsError as ex:
+                    if not overwrite: 
+                        _logger.info("El fichero existe, no sd sobreescribe por que overwrite=False")
+                        continue
+                    _logger.debug("El fichero existe, hay que sobreescribirlo")
+                    blob.delete_blob()
+                    _logger.debug("El fichero se ha borrado para sobreescribirlo")
+                    record._do_upload(blob, file_full_path)
 
-        if remove_local_data_file: os.unlink(self.file)
+                if remove_local_data_file: os.unlink(file_full_path)
+            except Exception as ex:
+                _logger.error("Error uploading to Blob: type({}), {}".format(type(ex), ex))            
+                continue
